+++ src/lem/eval.rs
@@ -24,8 +24,7 @@ pub(crate) fn step() -> Result<LEM> {
 mod tests {
     use super::*;
     use crate::field::LurkField;
-    use crate::lem::constrainer::AllocationManager;
-    use crate::lem::NumSlots;
+    use crate::lem::constrainer::{AllocationManager, NumSlots};
     use crate::lem::{pointers::Ptr, store::Store};
     use bellperson::util_cs::{test_cs::TestConstraintSystem, Comparable};
     use blstrs::Scalar as Fr;
